Rename chiplib to just "library"
