+++ octopus2/library/main.go
@@ -1,4 +1,4 @@
-package chiplib
+package library
 
 import "embed"
